WIP: Add RHCOS oscontainer into payload, render to 00-$role-osimageurl MC
This closes the gap in getting the RHCOS oscontainer into the update
payload, and having the MCO then render that down into the MachineConfigs.

Introduce a new ConfigMap `machine-config-osimageurl` that points to
the oscontainer and is applied by the CVO.  Then, the controller
renders that into `machineconfigs/00-$role-osimageurl` which then finally
goes into the rendered config, and should be applied by the daemon.

However, until we have CI side integration for managing the oscontainer,
add a new `CONFIG_IGNORE_OSIMAGEURL` environment variable option for
the daemon, and enable it by default.

Closes: https://github.com/openshift/machine-config-operator/issues/183