Add code to inject OSImageURL, but disable it by default
First, this supports injecting the `OSImageURL` into the "base"
config (e.g. `00-worker`, `00-master`).  This differs from
previous pull requests which made it a separate MC, but that
adds visual noise and will exacerbate renderer race conditions.

However, in order to gain experience with this code, add a
`ControllerConfig` option which can disable injecting it for certain
roles.  This is set to `*` by default, so effectively we won't
do OS updates.

My idea here is that anyone who wants to test things out can
`oc edit controllerconfig` and empty that out.  Another useful
thing would be to change it to e.g. `{"master"}` and test OS updates
on workers without affecting the master.