daemon: Move initial node bits more into checkStateOnFirstRun
Previously `getNodeAnnotation` read the initial annotations file,
and `getStateAndConfigs` also checked for its existence again.
Now we have a more clearly named `ensureNodeAnnotations` method
that's close to the other code handling the early state, and
it returns a `bool` denoting whether or not we found the initial
annotations file that's passed down into `getStateAndConfigs`.